requirements.txt
pandas==0.25.1
statsmodels==0.10.1
tqdm==4.28.1
hyperopt==0.2.3
numba==0.45.1
numpy==1.16.5
plotly==4.6.0
optuna==1.3.0
requests==2.22.0
matplotlib==3.1.3
sympy==1.4
networkx==2.4
scikit_learn==0.22.2.post1
scipy==1.4.1
streamlit
pyyaml
xlrd
